I forgot to 'forget' .user file also when I did the .suo file. It is note worthy that Visual Studio was showing wrong debugging information concerning variables when stepping in through the code. Deleting the suo file did the trick. I also searched for this problem and I found a blog entry talking about this: http://briannoyes.net/2008/05/10/the-evil-suo-file-fighting-and-winning-with-visual-studio/
Also, according to msdn, suo files should not be included in source control. Also
it is a binary file that always change, so the delta of changes can increase the
repo size overtime exponentially.

http://msdn.microsoft.com/en-us/library/bb165909.aspx

And from stackoverflow they suggested the exclusion of .user settings file.
Although it is a text file, and doesn't contain any real useful data, at least to me.

This commit is just a reminder for future references to me. Again, I am new to
Mercurial/Git repositories and version control.