- Moved the wxTimer derived classes into TwinPad_Frame, so when the dialogbox gets deleted, the timers will get deleted too since TwinPad_Frame is the Owner object. Also, I created a new timer for a planned Loading Combo Key, instead of using "Space bar" as default, to give the user more freedom.
- Note: I forgot to mention in the two commits, that I changed the solution a
bit, now there is a new build target, called "Run test with PCSX2". This is
meant only for testing and needs to be modified when tried on different computer
by changing the working directory of PCSX2-DBG.EXE and PLUGINS directories.

- Note 2: Just a reminder about debugging DLLs, we need to generate PDB symbol
files for the DLL and (if available) perhaps the Emu's. You get those by
modifying the solution to generate Debug info and specifying the path fo PDB
files. Very important, you have to keep the symbols files in the same directory
as the working directory and plugins directory. That is, the generated DLL and
PDB files in the same directory, and the EXE file (from the Emu if available)
with its PDB file.. By doing this, you will be able to set Breakpoints and
step through the code while running. Which makes life easier for debugging DLLs.