- Reverted one line that caused grid to crash when doing custom cell rendering. When I removed wxInitAllImageHandlers() the GUI successfuly loaded animated GIFs. But, without this, loading GIFs as BMPs in custom cell renderer fails and causes an exception. Fixed.
- Implemented GetKeyForCombo() to retrieve a key and assign it to a combo.
Fixed the code to support saving and loading combos with the configured key while
switching combos from the combobox.

TODO: Add GetKeyForMisc() for the misc tab. And assign that as a hot key to show
the config window on the fly, edit combos and reload configs while emu is running.