Misc fixes: - Removed another memory leak: wxGridCellAttr should not be IncRef() manually, instead create a new wxGridCellAttr for each column instead of sharing one for all of them. Delete operator doesn't work with Reference Counted, thus it needed DecRef() before. - Added macro to detect memory leaks by redefining the new operator. Thanks to MSDN website. - Pumped TwinPad version to v0.9.2 :)
Note: There might be a false positive memory leak (74 bytes) during
exit. If the user "configures" twinpad without running the emulation
and exit the emu. (Happens only with PCSX2 and not PCSXR).
And this happens even if I don't use twinpad at all and not even
loaded, so it must be from PCSX2 or the other plugins. I could
be wrong though. If the user runs the emulation nothing shows up
in the debug output window. This could be also a false positive from
using a global class (not a singleton, but similar), and the leak is
detected before cleaning the global variables.